Renaming package reflect to invoke. Closes #191
