+++ vraptor-core/src/main/java/br/com/caelum/vraptor/interceptor/ExecuteMethodInterceptor.java
@@ -30,8 +30,8 @@ import br.com.caelum.vraptor.Validator;
 import br.com.caelum.vraptor.controller.ControllerMethod;
 import br.com.caelum.vraptor.core.InterceptorStack;
 import br.com.caelum.vraptor.core.MethodInfo;
-import br.com.caelum.vraptor.reflection.MethodExecutor;
-import br.com.caelum.vraptor.reflection.MethodExecutorException;
+import br.com.caelum.vraptor.invoke.MethodExecutor;
+import br.com.caelum.vraptor.invoke.MethodExecutorException;
 import br.com.caelum.vraptor.util.Stringnifier;
 import br.com.caelum.vraptor.validator.ValidationException;
 import static br.com.caelum.vraptor.view.Results.nothing;
