+++ vraptor-core/src/main/java/br/com/caelum/vraptor/interceptor/StepInvoker.java
@@ -10,8 +10,8 @@ import net.vidageek.mirror.dsl.Mirror;
 import net.vidageek.mirror.list.dsl.Matcher;
 import net.vidageek.mirror.list.dsl.MirrorList;
 import br.com.caelum.vraptor.InterceptionException;
-import br.com.caelum.vraptor.reflection.MethodExecutor;
-import br.com.caelum.vraptor.reflection.MethodExecutorException;
+import br.com.caelum.vraptor.invoke.MethodExecutor;
+import br.com.caelum.vraptor.invoke.MethodExecutorException;
 
 import com.google.common.base.Throwables;
