+++ vraptor-core/src/main/java/br/com/caelum/vraptor/invoke/DefaultMethodExecutor.java
@@ -1,4 +1,4 @@
-package br.com.caelum.vraptor.reflection;
+package br.com.caelum.vraptor.invoke;
 
 import java.lang.invoke.MethodHandle;
 import java.lang.reflect.Method;
