+++ vraptor-core/src/main/java/br/com/caelum/vraptor/invoke/MethodExecutor.java
@@ -1,4 +1,4 @@
-package br.com.caelum.vraptor.reflection;
+package br.com.caelum.vraptor.invoke;
 
 import java.lang.reflect.Method;
