+++ vraptor-core/src/main/java/br/com/caelum/vraptor/invoke/MethodHandleFactory.java
@@ -1,4 +1,4 @@
-package br.com.caelum.vraptor.reflection;
+package br.com.caelum.vraptor.invoke;
 
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
