+++ vraptor-core/src/test/java/br/com/caelum/vraptor/interceptor/ExecuteMethodInterceptorTest.java
@@ -19,12 +19,9 @@ package br.com.caelum.vraptor.interceptor;
 
 import static br.com.caelum.vraptor.view.Results.nothing;
 import static org.hamcrest.MatcherAssert.assertThat;
-
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
-
 import static org.junit.Assert.assertTrue;
-
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -47,7 +44,7 @@ import br.com.caelum.vraptor.controller.DefaultControllerMethod;
 import br.com.caelum.vraptor.core.InterceptorStack;
 import br.com.caelum.vraptor.core.MethodInfo;
 import br.com.caelum.vraptor.factory.Factories;
-import br.com.caelum.vraptor.reflection.MethodExecutor;
+import br.com.caelum.vraptor.invoke.MethodExecutor;
 import br.com.caelum.vraptor.validator.Message;
 import br.com.caelum.vraptor.validator.ValidationException;
